requirements.txt
google-genai>=0.8.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
schedule>=1.2.1
streamlit>=1.35.0
